SS-93 Refactor exceptions folder
