+++ backend/src/main/java/ru/hh/superscoring/util/exceptions/TestPassTimeoutException.java
@@ -1,4 +1,4 @@
-package ru.hh.superscoring.util.Exceptions;
+package ru.hh.superscoring.util.exceptions;
 
 public class TestPassTimeoutException extends Exception {
   public TestPassTimeoutException(String errorMessage) {
